Final/blog (#5)
* refactor(blog): Rename blogData to blogItemData and update references for clarity

- Change variable name from blogData to blogItemData in blog-data.ts for consistency.
- Update BlogListRenderer to use blogItemData in createBlogItemElement and render methods.
- Adjust import statements in blog.ts to reflect the new variable name.

* refactor(blog): Allow customizable selector for BlogListRenderer

- Update BlogListRenderer constructor to accept a selector parameter, defaulting to ".blog-list".
- Add a comment in blog.ts to indicate the rendering of the blog list.

* feat(blog): Implement BlogListRenderer for dynamic blog item rendering

- Create BlogListRenderer class to handle rendering of blog items in the DOM.
- Add methods for generating HTML for blog items, tags, and popular tags.
- Update blog.ts to import BlogListRenderer from the new components directory.

* refactor(blog): Update blog layout and rendering logic

- Replace the previous content structure in blog.html with a new header and navigation.
- Rename the content class to main-content in SCSS for consistency with the updated HTML.
- Modify BlogListRenderer to return the rendered blog list element instead of directly manipulating the DOM.
- Update blog.ts to append the rendered blog list to the main-content area.

* refactor(blog): Remove header and navigation from blog layout

- Eliminate the header and primary navigation elements from blog.html to simplify the layout.
- Maintain the main content structure for future enhancements.
