+++ main-pset/final-project/blog.html
@@ -11,9 +11,7 @@
   </head>
   <body>
     <div class="wrap">
-      <div class="content">
-        <ul class="blog-list"></ul>
-      </div>
+      <div class="main-content"></div>
     </div>
   </body>
 </html>
