+++ main-pset/final-project/src/scss/_layout.scss
@@ -6,8 +6,8 @@
   justify-content: center;
 }
 
-.content {
-  max-width: $max-width; 
+.main-content {
+  max-width: $max-width;
   display: flex;
   justify-content: center;
   margin: 0 10px;
